Optional default values from environment variables
Tags `env:"VARIABLE_NAME"` and `env-delim:"DELIM_STRING"` can
be optionally specified to get the option default value from
the corresponding environment variable.

The delimiter, if one is defined, is used to split the value
to multiple strings, making it possible to define default values
for both slices and maps thru the environment.